refactor: move List to Molecules
